Changing screen name from configuration to profile
